src/main/java/com/example/JavaFinalProject/service/ProductService.java
package com.example.JavaFinalProject.service;

import com.example.JavaFinalProject.dtos.ProductDto;
import com.example.JavaFinalProject.model.Product;
import com.example.JavaFinalProject.repository.ProductRepository;
import jakarta.persistence.EntityNotFoundException;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;
import java.util.Optional;
import java.util.stream.Collectors;


@Service
public class ProductService implements ProductServiceInterface {

    @Autowired
    private ProductRepository productRepository;

    // Constructor care primește un ProductRepository
    public ProductService(ProductRepository productRepository) {
        this.productRepository = productRepository;
    }
    @Override
    public void addProduct(ProductDto productDto) {
        if (productDto.getName() == null || productDto.getName().isEmpty()) {
            throw new IllegalArgumentException("Product name cannot be null or empty!");
        }
        if (productDto.getPrice() == null || productDto.getPrice().doubleValue() < 0) {
            throw new IllegalArgumentException("Product price cannot be null or negative!");
        }

        Product product = new Product(
                productDto.getName(),
                productDto.getDescription(),
                productDto.getPrice(),
                productDto.getCategory(),
                10 // Valoare implicită pentru stoc
        );

        productRepository.save(product);
    }

    @Override
    public List<ProductDto> getAllProducts() {
        return productRepository.findAll().stream()
                .map(product -> new ProductDto(
                        product.getId(),
                        product.getName(),
                        product.getDescription(),
                        product.getPrice(),
                        product.getCategory()))
                .collect(Collectors.toList());
    }

    @Override
    public Product findProductById(Long id) {
        return productRepository.findById(id)
                .orElseThrow(() -> new EntityNotFoundException("Product with ID " + id + " not found!"));
    }



    @Override
    public void updateProductStock(Long id, Integer quantity) {
        Product product = productRepository.findById(id)
                .orElseThrow(() -> new EntityNotFoundException("Product with ID " + id + " not found!"));

        product.setStockQuantity(quantity);
        productRepository.save(product);
    }

    @Override
    public boolean updateProduct(Long id, ProductDto productDto) {
        Optional<Product> existingProduct = productRepository.findById(id);

        if (existingProduct.isPresent()) {
            Product product = existingProduct.get();

            if (productDto.getName() != null && !productDto.getName().isEmpty()) {
                product.setName(productDto.getName());
            }
            if (productDto.getDescription() != null) {
                product.setDescription(productDto.getDescription());
            }
            if (productDto.getPrice() != null && productDto.getPrice().doubleValue() >= 0) {
                product.setPrice(productDto.getPrice());
            }
            if (productDto.getCategory() != null) {
                product.setCategory(productDto.getCategory());
            }

            productRepository.save(product);
            return true;
        }
        return false;
    }

    @Override
    public ProductDto getProductById(Long id) {
        Product product = productRepository.findById(id)
                .orElseThrow(() -> new EntityNotFoundException("Product with ID " + id + " not found!"));

        return new ProductDto(
                product.getId(),
                product.getName(),
                product.getDescription(),
                product.getPrice(),
                product.getCategory());
    }



    @Override
    public boolean deleteProduct(Long id) {
        if (productRepository.existsById(id)) {
            productRepository.deleteById(id);
            return true;
        }
        return false;
    }


}
